Migrate library to the new framework version.
