+++ datastore/src/main/java/io/spine/server/storage/datastore/DatastoreWrapper.java
@@ -34,7 +34,6 @@ import com.google.cloud.datastore.KeyFactory;
 import com.google.cloud.datastore.Query;
 import com.google.cloud.datastore.StructuredQuery;
 import com.google.cloud.datastore.Transaction;
-import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Supplier;
 import com.google.common.collect.Streams;
 import io.spine.logging.Logging;
@@ -496,7 +495,6 @@ public class DatastoreWrapper implements Logging {
         return options;
     }
 
-    @VisibleForTesting
     Datastore getDatastore() {
         return datastore;
     }
