+++ datastore/src/main/java/io/spine/server/storage/datastore/DsLookupByIds.java
@@ -38,7 +38,6 @@ import java.util.stream.Stream;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.collect.Streams.stream;
-import static io.spine.server.entity.EntityWithLifecycle.Predicates.isRecordActive;
 import static io.spine.server.storage.datastore.DsEntityComparator.implementing;
 import static io.spine.server.storage.datastore.Entities.toRecord;
 import static io.spine.server.storage.datastore.FieldMaskApplier.nullableRecordMasker;
@@ -203,7 +202,8 @@ final class DsLookupByIds<I> {
     private Stream<EntityRecord> readActiveRecords(Iterable<I> ids) {
         Stream<EntityRecord> records = read(ids)
                 .map(Entities::nullableToRecord)
-                .map(nullIfNot(isRecordActive()));
+                .filter(Objects::nonNull)
+                .map(nullIfNot(EntityRecord::isActive));
         return records;
     }
