+++ datastore/src/main/java/io/spine/server/storage/datastore/tenant/NamespaceSupplier.java
@@ -23,9 +23,9 @@ package io.spine.server.storage.datastore.tenant;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Supplier;
 import io.spine.annotation.Internal;
+import io.spine.core.TenantId;
 import io.spine.server.storage.datastore.DatastoreStorageFactory;
 import io.spine.server.storage.datastore.ProjectId;
-import io.spine.core.TenantId;
 
 import javax.annotation.Nullable;
 
@@ -59,7 +59,8 @@ public abstract class NamespaceSupplier implements Supplier<Namespace> {
         }
     }
 
-    public static NamespaceSupplier singleTenant() {
+    @VisibleForTesting
+    static NamespaceSupplier singleTenant() {
         return new SingleTenantNamespaceSupplier(null);
     }
