+++ datastore/src/test/java/io/spine/server/storage/datastore/DsAggregateStorageTest.java
@@ -35,7 +35,7 @@ import io.spine.server.aggregate.AggregateReadRequest;
 import io.spine.server.aggregate.AggregateStorage;
 import io.spine.server.aggregate.AggregateStorageTest;
 import io.spine.server.aggregate.Snapshot;
-import io.spine.server.aggregate.given.AggregateRepositoryTestEnv.ProjectAggregate;
+import io.spine.server.aggregate.given.repo.ProjectAggregate;
 import io.spine.server.entity.Entity;
 import io.spine.server.storage.datastore.given.DsAggregateStorageTestEnv.NonProjectStateAggregate;
 import io.spine.server.storage.datastore.given.aggregate.ProjectAggregateRepository;
@@ -173,7 +173,7 @@ class DsAggregateStorageTest extends AggregateStorageTest {
 
     @Test
     @DisplayName("not overwrite the snapshot when saving a new one")
-    void notOverwriteSnapshot() throws InterruptedException {
+    void notOverwriteSnapshot() {
         DsAggregateStorage<ProjectId> storage = (DsAggregateStorage<ProjectId>) getStorage();
         ProjectId id = newId();
